better spot for iter
